Rework the navigation.
The navigation is now a component, although it is only used once. This
keeps things neat and tidy and allows the JS responsible for toggling
the navigation to be kept separate.

Also modified the way the push happens. Have gone back to a transform,
except now the topbar and the navigation appear outside the container
so that we don't get that weird fixed positioning that occurs
when transforming a fixed element.