Begin working on user authentication.
Started the basics of user authentication. Pulled in the JWT library
and configured some database odds and ends.

Have made some changes to the default AuthController to better allow
it to respond to AJAX.

Decided against pulling in Dingo for now, will roll a simplified
custom solution for the API considering there will be no actual
Laravel routes aside from the default route to load things.

Also created a new Alert and Error Alert component that work
in tandum to display alerts, and more importantly, errors from
when requests fail.